Clean up and documentation.
